Move Compat.TempFile to Internal.* module and expose from Cabal
This is required for the reworking of integration tests
to use shell scripts.
